+++ cli/cli/command_framework/highlevel/engine_consuming_kurtosis_command/engine_consuming_kurtosis_command.go
@@ -16,9 +16,6 @@ import (
 
 const (
 	engineClientCloseFuncCtxKey = "engine-client-close-func"
-
-	// TODO Delete this and instead replace with reading cluster name from disk!
-	clusterName = "docker"
 )
 
 // This is a convenience KurtosisCommand for commands that interact with the engine
@@ -105,9 +102,9 @@ func (cmd *EngineConsumingKurtosisCommand) getSetupFunc() func(context.Context) 
 	return func(ctx context.Context) (context.Context, error) {
 		result := ctx
 
-		engineManager, err := engine_manager.NewEngineManager(clusterName)
+		engineManager, err := engine_manager.NewEngineManager()
 		if err != nil {
-			return nil, stacktrace.Propagate(err, "An error occurred getting an engine manager connected to cluster '%v'", clusterName)
+			return nil, stacktrace.Propagate(err, "An error occurred getting an engine manager.")
 		}
 
 		// TODO This is a hack that's only here temporarily because we have commands that use KurtosisBackend directly (they