+++ cli/cli/commands/module/exec/exec.go
@@ -55,9 +55,6 @@ const (
 	netReadOpt = "read"
 
 	netReadOptFailBecauseSourceIsUsedOrClosedErrorText = "use of closed network connection"
-
-	// TODO remove when we read cluster state from disk
-	clusterName = "docker"
 )
 
 var positionalArgs = []string{
@@ -160,9 +157,9 @@ func run(cmd *cobra.Command, args []string) error {
 		)
 	}
 
-	engineManager, err := engine_manager.NewEngineManager(clusterName)
+	engineManager, err := engine_manager.NewEngineManager()
 	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred creating an engine manager connected to cluster '%v'", clusterName)
+		return stacktrace.Propagate(err, "An error occurred creating an engine manager.")
 	}
 	// TODO THIS IS A BIG JANKY HACK. We should instead migrate this commmand to be an EngineConsumingKurtosisCommand instead
 	kurtosisBackend := engineManager.GetKurtosisBackend()